Added user-defined directory functionality
I know this is a 2-year-old script but I found it and thought it could do a little bit more so I added the ability to set the location for not only where the CIA/3DS files are but the place where the decrypted file is placed. I did this by making the script copy over all the necessary exes and change the working directory to the user set one. The user set preferences are saved in "locations.txt" so that subsequent runs don't require the user's input and remember where the user stored their files. I added that, during the deletion of the ncch and decfirst files, it also deletes the exes so there is nothing left besides the log and the decrypted file. Everything is still left intact in the script's home directory. The other changes are related, like that there is a check to make sure the file path actually exists as well as checking for it to be in the right syntax. Honestly, this works for my specific situation since I store this script separate from my 3ds/cia files so it might not be useful to anyone else but I figured I'd put my changes up just in case. I have more ideas on stuff I could add but I feel like that would just be for my own use and my need to overcomplicate things that don't need to be.